nav bar implemented, radial bar mediaquery, systemchrome
