+++ lib/main.dart
@@ -1,7 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:soil_moisture_app/ui/build_theme.dart';
-import 'package:percent_indicator/circular_percent_indicator.dart';
-import 'package:soil_moisture_app/ui/plant_card.dart';
+import 'package:soil_moisture_app/ui/colors.dart';
+import 'package:flutter/services.dart';
+// Pages Import
+
+import 'pages/Analysis.dart';
+import 'pages/Overview.dart';
 
 import 'dart:math' as math; //! Remove this when refresh implemented
 
@@ -27,171 +31,65 @@ class HomeApp extends StatefulWidget {
 }
 
 class _HomeAppState extends State<HomeApp> {
-  final GlobalKey<RefreshIndicatorState> _refreshIndicatorKey =
-      new GlobalKey<RefreshIndicatorState>();
-  int _counter = 0;
-
-// * Dummy pull to refresh implemented, remove when done
-  Future<void> dummyWait() {
-    return Future.delayed(Duration(seconds: 5), null);
-  }
-
-  Future<Null> _refresh() {
-    return dummyWait().then((_) {
-      setState(() => _counter = rnd.nextInt(101));
-    });
-  }
-
+  int _currentIndex = 0;
+  final List<Widget> _children = [
+    Overview(),
+    Analysis(),
+  ];
   @override
   Widget build(BuildContext context) {
+    SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle(
+      statusBarColor: Colors.black, //top bar color
+      systemNavigationBarColor: Colors.black, //bottom bar color
+
+      statusBarIconBrightness: Brightness.light, //top icon color
+      systemNavigationBarIconBrightness: Brightness.light, //bottom icons color
+    ));
     return Scaffold(
       appBar: AppBar(
         title: Text(widget.title),
         centerTitle: true,
       ),
-      body: RefreshIndicator(
-        key: _refreshIndicatorKey,
-        onRefresh: _refresh,
-        child: SafeArea(
-          child: Padding(
-            padding: const EdgeInsets.only(left: 15.0, right: 15.0),
-            child: ListView(
-              physics: BouncingScrollPhysics(),
-              children: <Widget>[
-                SizedBox(
-                  height: 5.0,
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceAround,
-                  children: <Widget>[
-                    CircularPercentIndicator(
-                      animationDuration: 600,
-                      radius: 300.0,
-                      animation: true,
-                      percent: _counter / 100,
-                      circularStrokeCap: CircularStrokeCap.round,
-                      backgroundColor: Colors.grey[300],
-                      progressColor:
-                          (_counter < 15) ? Colors.red : Colors.green,
-                      lineWidth: 10.0,
-                      center: Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: <Widget>[
-                          Text(
-                            '$_counter%',
-                            style:
-                                Theme.of(context).textTheme.display4.copyWith(
-                                      fontSize: 130.0,
-                                    ),
-                          ),
-                          Text(
-                            'Current Moisture',
-                            style:
-                                Theme.of(context).textTheme.display1.copyWith(
-                                      fontSize: 24.0,
-                                    ),
-                          )
-                        ],
-                      ),
-                    ),
-                    Container(
-                      height: 150.0,
-                      width: 150.0,
-                      child: Placeholder(),
-                    ),
-                  ],
-                ),
-                SizedBox(
-                  height: 20.0,
-                ),
-                Card(
-                  child: ListTile(
-                    title: Text(
-                      'You have pushed the button this many times:(Some Info Here)',
-                    ),
-                    subtitle: Text(
-                      '$_counter',
-                      style: Theme.of(context).textTheme.display1,
-                    ),
-                  ),
-                ),
-                SizedBox(
-                  height: 20.0,
-                ),
-                Wrap(
-                  alignment: WrapAlignment.center,
-                  runSpacing: 5.0,
-                  children: <Widget>[
-                    PlantCard(
-                      title: 'Plant 1',
-                      img: '',
-                      onTap: () => print('Plant1'),
-                      crit: 0.15,
-                      percent: _counter/100,
-                    ),
-                    PlantCard(
-                      title: 'Plant 2',
-                      img: '',
-                    ),
-                    PlantCard(
-                      title: 'Plant 3',
-                      img: '',
-                    ),
-                    PlantCard(
-                      title: 'Plant 4',
-                      img: '',
-                    ),
-                    PlantCard(
-                      title: 'Plant 5',
-                      img: '',
-                    ),
-                    PlantCard(
-                      title: 'Plant 6',
-                      img: '',
-                    ),
-                    PlantCard(
-                      title: 'Plant 7',
-                      img: '',
-                    ),
-                    PlantCard(
-                      title: 'Plant 8',
-                      img: '',
-                    ),
-                  ],
-                )
-              ],
-            ),
-          ),
-        ),
-      ),
+      body: _children[_currentIndex],
       bottomNavigationBar: new BottomNavigationBar(
+        onTap: onTabTapped,
         type: BottomNavigationBarType.fixed,
         currentIndex: 0,
         items: [
           BottomNavigationBarItem(
             icon: Icon(
               Icons.remove_red_eye,
+              color: _currentIndex == 0 ? appSecondaryDarkColor : appPrimaryDarkColor,
             ),
             title: new Text(
               "Overview",
               style: TextStyle(
                 fontFamily: 'Ocrb',
+                color: _currentIndex == 0 ? appSecondaryDarkColor : appPrimaryDarkColor,
               ),
             ),
           ),
           BottomNavigationBarItem(
             icon: Icon(
               Icons.linear_scale,
+              color: _currentIndex == 1 ? appSecondaryDarkColor : appPrimaryDarkColor,
             ),
             title: new Text(
               "Analysis",
               style: TextStyle(
                 fontFamily: 'Ocrb',
+                color: _currentIndex == 1 ? appSecondaryDarkColor : appPrimaryDarkColor,
               ),
             ),
           ),
         ],
       ),
     );
+  }
+
+  void onTabTapped(int index) {
+    setState(() {
+      _currentIndex = index;
+    });
   }
 }
