+++ lib/pages/Analysis.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Analysis extends StatefulWidget {
+  @override
+  _AnalysisState createState() => _AnalysisState();
+}
+
+class _AnalysisState extends State<Analysis> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(child: Text("Analysis")),
+    );
+  }
+}
