+++ lib/ui/plant_card.dart
@@ -14,11 +14,11 @@ class PlantCard extends StatelessWidget {
       this.img,
       this.onTap,
       this.percent = 0.15,
-      this.crit = 0.15}){
-        if(this.onTap == null){
-          this.onTap = () => print(this.title);
-        }
-      }
+      this.crit = 0.15}) {
+    if (this.onTap == null) {
+      this.onTap = () => print(this.title);
+    }
+  }
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -39,10 +39,10 @@ class PlantCard extends StatelessWidget {
               children: <Widget>[
                 Text(
                   this.title,
-                  style: Theme.of(context).textTheme.body2.copyWith(
-                    fontFamily: 'Ocrb',
-                    fontSize: 18.0
-                  ),
+                  style: Theme.of(context)
+                      .textTheme
+                      .body2
+                      .copyWith(fontFamily: 'Ocrb', fontSize: 18.0),
                 ),
                 Container(
                   height: 50.0,
